First_Trimester_Exams/requirements.txt
Faker==21.0.0
psycopg==3.1.15
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
typing_extensions==4.9.0
